requirements.txt
# Core PDF and data processing
reportlab==4.0.7
pandas==2.1.4
openpyxl==3.1.2

# Barcode generation (v1.x and v2.x)
python-barcode==0.16.1
qrcode==8.2

# Image processing
Pillow==10.1.0

# Date arithmetic (accurate month calculations)
python-dateutil==2.8.2

# DoD/NATO label requirements (v2.x)
pylibdmtx==0.1.10
segno==1.6.6
